+++ Changes.md
@@ -18,6 +18,7 @@ Improvements
   - Added `destination` plug, to control where the copies are placed relative to the
     original.
   - Improved performance for large numbers of copies.
+  - Deprecated the `target` plug. Please use filters instead.
 - Outputs : Reduced the time taken to show the NodeEditor by around 90%.
 - NodeEditor : The "Node Name" label is now draggable. For instance, it can be dragged to the PythonEditor to get a reference to the node or to the GraphEditor to find the node in the graph.
 - GraphEditor : Improved framing of nodes dragged and dropped onto the GraphEditor :